Modules/Python/requirements.txt
sqlalchemy~=1.4.26
python-chess
discord.py
PyNaCl
motor~=2.4.0
aiml~=0.9.2
gtts~=2.2.2
lyricsgenius~=3.0.1
spotipy~=2.17.1
mee6_py_api
mysql.connector
youtube-search-python
pytube~=10.7.1
youtube_dl
cairosvg (download https://downloads.sk1project.net/sk1/MS_Windows/sk1-2.0rc5-win64_headless.msi and set to sys env var)
cryptography~=35.0.0
chess~=1.7.0
CairoSVG~=2.5.2
mysql-connector-python~=8.0.24
aiohttp~=3.7.3
pip~=20.2.3
wheel~=0.37.0
Brotli~=1.0.9
lxml~=4.6.3
setuptools~=49.2.1
future~=0.18.2
pydub~=0.25.1
requests~=2.25.1